+++ views/awesome-support-required.php
@@ -1,0 +1,7 @@
+<div class="notice notice-error is-dismissible">
+    <p><?php
+        echo esc_html(
+            __('The Awesome Support: Importer requires Awesome Support to be activated.', 'awesome-support-importer')
+        );
+        ?></p>
+</div>